+++ src/collision.c
@@ -6,22 +6,22 @@
 /*   By: macastro <macastro@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/21 14:59:52 by macastro          #+#    #+#             */
-/*   Updated: 2025/03/03 17:37:27 by macastro         ###   ########.fr       */
+/*   Updated: 2025/03/05 12:47:27 by macastro         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "cub3d.h"
 
-t_bool	check_wall_collision(int point[2], int offset_x, int offset_y,
+t_bool	check_wall_collision(float point[2], int offset_x, int offset_y,
 	t_world *world)
 {
-	int	next_point[2];
+	float	next_point[2];
 
 	assign_point_ints(next_point, point[X] + offset_x, point[Y] + offset_y);
 	return (pos_is_wall(next_point, world));
 }
 
-t_bool	too_near_wall(int point[2], t_world *world)
+t_bool	too_near_wall(float point[2], t_world *world)
 {
 	const int	min_dist = CUB_SIZE / 4;
 	int			grid_pos[2];
@@ -29,8 +29,8 @@ t_bool	too_near_wall(int point[2], t_world *world)
 
 	grid_pos[R] = grid_row(point);
 	grid_pos[C] = grid_column(point);
-	cell_pt[X] = point[X] % CUB_SIZE;
-	cell_pt[Y] = point[Y] % CUB_SIZE;
+	cell_pt[X] = (int)point[X] % CUB_SIZE;
+	cell_pt[Y] = (int)point[Y] % CUB_SIZE;
 	if (cell_pt[X] < min_dist && check_wall_collision(point,
 			-cell_pt[X] - 1, 0, world)) // left
 		return (TRUE);
@@ -46,7 +46,7 @@ t_bool	too_near_wall(int point[2], t_world *world)
 	return (world->map[grid_pos[R]][grid_pos[C]] == '1');
 }
 
-t_bool	pos_is_wall(int point[2], t_world *world)
+t_bool	pos_is_wall(float point[2], t_world *world)
 {
 	int	grid_pos[2];
